+++ pymilo/pymilo_param.py
@@ -13,7 +13,8 @@ import sklearn.dummy as dummy
 import sklearn.ensemble as ensemble
 import sklearn.pipeline as pipeline
 import sklearn.preprocessing as preprocessing
-from sklearn.cross_decomposition import PLSRegression, PLSCanonical, CCA
+import sklearn.cross_decomposition as cross_decomposition
+import sklearn.feature_extraction as feature_extraction
 
 quantile_regressor_support = False
 try:
@@ -246,10 +247,25 @@ SKLEARN_PREPROCESSING_TABLE = {
     "TargetEncoder": TargetEncoder if target_encoder_support else NOT_SUPPORTED,
 }
 
+SKLEARN_FEATURE_EXTRACTION_TABLE = {
+    # for raw data:
+    "DictVectorizer": feature_extraction.DictVectorizer,
+    "FeatureHasher": feature_extraction.FeatureHasher,
+
+    # for image data:
+    "PatchExtractor": feature_extraction.image.PatchExtractor,
+
+    # for text data:
+    "CountVectorizer": feature_extraction.text.CountVectorizer,
+    "HashingVectorizer": feature_extraction.text.HashingVectorizer,
+    "TfidfTransformer": feature_extraction.text.TfidfTransformer,
+    "TfidfVectorizer": feature_extraction.text.TfidfVectorizer,
+}
+
 SKLEARN_CROSS_DECOMPOSITION_TABLE = {
-    "PLSRegression": PLSRegression,
-    "PLSCanonical": PLSCanonical,
-    "CCA": CCA,
+    "PLSRegression": cross_decomposition.PLSRegression,
+    "PLSCanonical": cross_decomposition.PLSCanonical,
+    "CCA": cross_decomposition.CCA,
 }
 
 KEYS_NEED_PREPROCESSING_BEFORE_DESERIALIZATION = {